Sweep the (s16)-on-s16 no-op cast class
Repo-wide census found 48 candidate casts of s16 variables to their own
width. Droppable (matchdiff MATCH): draw_time_ (6), draw_digits_ (2),
DrawConstruction (8), ActivateHumans (6), DamageControl's (short)dmg/hp
cluster (12, note quotes updated), AddEnemy (plus 0x31 spelled '1'),
ControlTraceLine (plus the register-named a0 local renamed diff).
Byte-required and reverted: mission_score_screen and return_to_menu_ —
there the variable was last written from a wider expression, so the
cast materializes the sll/sra re-extension. briefing_screen_ and the
Briefing screen's lever sites are measured territory and untouched.
Mechanism recorded in the cookbook. Whole image byte-identical.

+++ src/main.exe/ActivateHumans.c
@@ -263,16 +263,16 @@ void ActivateHumans(void)
             human->type == GOO)
         {
             j = 0;
-            while (StageChar[(s16)j].stage != -1)
+            while (StageChar[j].stage != -1)
             {
                 do
                 {
-                    if (StageChar[(s16)j].stage == StageID + 1 &&
-                        StageChar[(s16)j].chrid == human->type)
+                    if (StageChar[j].stage == StageID + 1 &&
+                        StageChar[j].chrid == human->type)
                     {
-                        human->model->locate.coord.t[0] = StageChar[(s16)j].position.vx * 1000;
-                        human->model->locate.coord.t[1] = StageChar[(s16)j].position.vy * 1000;
-                        human->model->locate.coord.t[2] = StageChar[(s16)j].position.vz * 1000;
+                        human->model->locate.coord.t[0] = StageChar[j].position.vx * 1000;
+                        human->model->locate.coord.t[1] = StageChar[j].position.vy * 1000;
+                        human->model->locate.coord.t[2] = StageChar[j].position.vz * 1000;
                     }
                 } while (0);
                 j++;
